convert-mme-json-to-csv/requirements.txt
functions-framework==3.*
pandas
google-cloud-storage
fsspec
gcsfs
